feat(mcp): extract submission helpers into submissions-lib
Move MCP submission parsing, validation, duplicate search, and draft
builders into a dedicated lib module with wrapper re-exports and tests.
